+++ src/constants.ts
@@ -1,4 +1,4 @@
-export enum CommandCategory {
+export const enum CommandCategory {
     SUPPORT = 'Support',
     GENERAL = 'General',
     UTILS = 'Utilities',
@@ -6,7 +6,7 @@ export enum CommandCategory {
     MODERATION = 'Moderation',
 }
 
-export enum ResponseType {
+export const enum ResponseType {
     STATIC,
     EPHEMERAL,
     EDIT,
@@ -14,7 +14,7 @@ export enum ResponseType {
     FOLLOW_UP,
 }
 
-export enum OpenAIModels {
+export const enum OpenAIModels {
     GPT4 = 'gpt-4',
     GPT4_32K = 'gpt-4-32k',
     GPT3_TURBO = 'gpt-3.5-turbo',
@@ -23,4 +23,8 @@ export enum OpenAIModels {
     CODE_CUSHMAN = 'code-cushman-001',
     TEXT_DAVINCI_EDIT = 'text-davinci-edit-001',
     CODE_DAVINCI_EDIT = 'code-davinci-edit-001',
+}
+
+export const enum MessageID {
+    ROLE_SELECT = '1097532966835605605',
 }