+++ .github/CHANGELOG.md
@@ -1,3 +1,6 @@
+# April 2023
+- feat: (04/18/2023): Add new role handler & command
+
 # March 2023
 - feat: (03/07/2023): Integrate slash command api & new cowsay command
 - feat: (03/27/2023): Integrate OpenAI API & ChatGPT Command